Move into the correct folder
