--- result-coroutines/src/main/kotlin/com/github/kittinunf/result/coroutines/SuspendableValidation.kt
@@ -1,9 +1,0 @@
-package com.github.kittinunf.result.coroutines
-
-class SuspendedValidation<out E : Exception>(vararg resultSequence: SuspendableResult<*, E>) {
-
-    val failures: List<E> = resultSequence.filterIsInstance<SuspendableResult.Failure<*, E>>().map { it.getException() }
-
-    val hasFailure = failures.isNotEmpty()
-
-}
