+++ result/build.gradle.kts
@@ -1,15 +1,91 @@
-sourceSets {
-    getByName("main").java.srcDirs("src/main/kotlin")
-    getByName("test").java.srcDirs("src/main/kotlin")
+import java.io.IOException
+
+plugins {
+    kotlin("multiplatform")
+    java
+    jacoco
+    id("publication")
 }
 
-dependencies {
-    val junit: String by project
-    val kotlinVersion: String by project
-    val coroutines: String by project
+val artifactGroupId: String by project
+group = artifactGroupId
 
-    implementation(kotlin("stdlib", kotlinVersion))
+val gitSha = "git rev-parse --short HEAD".runCommand(project.rootDir)?.trim().orEmpty()
+val artifactPublishVersion: String by project
 
-    testImplementation("junit:junit:$junit")
-    testImplementation("org.jetbrains.kotlinx:kotlinx-coroutines-core:$coroutines")
+val isReleaseBuild: Boolean
+    get() = properties.containsKey("release")
+
+version = if (isReleaseBuild) artifactPublishVersion else "master-$gitSha-SNAPSHOT"
+
+kotlin {
+    jvm()
+
+    sourceSets {
+        all {}
+
+        val commonMain by getting {
+            dependencies {
+                implementation(libs.kotlin.stdlib)
+            }
+        }
+
+        val commonTest by getting {
+            dependencies {
+                implementation(libs.bundles.kotlin.test)
+            }
+        }
+
+        val jvmMain by getting {}
+
+        val jvmTest by getting {
+            dependencies {
+                implementation(libs.kotlin.test.junit)
+            }
+        }
+    }
+}
+
+jacoco {
+    toolVersion = libs.versions.jacoco.get()
+}
+
+tasks {
+
+    withType<JacocoReport> {
+        group = "Reporting"
+        description = "Generate Jacoco coverage reports."
+
+        val jvmTest by getting
+        dependsOn(jvmTest)
+
+        val classFiles = File("$buildDir/classes/kotlin/jvm/main").walkBottomUp().toSet()
+        classDirectories.setFrom(classFiles)
+        sourceDirectories.setFrom(files(arrayOf("$projectDir/src/commonMain")))
+        executionData.setFrom(files("$buildDir/jacoco/jvmTest.exec"))
+
+        reports {
+            xml.required.set(true)
+
+            html.required.set(true)
+            html.outputLocation.set(buildDir.resolve("reports"))
+
+            csv.required.set(false)
+        }
+    }
+}
+
+fun String.runCommand(workingDir: File): String? = try {
+    val parts = split("\\s".toRegex())
+    val proc = ProcessBuilder(*parts.toTypedArray())
+        .directory(workingDir)
+        .redirectOutput(ProcessBuilder.Redirect.PIPE)
+        .redirectError(ProcessBuilder.Redirect.PIPE)
+        .start()
+
+    proc.waitFor(30, TimeUnit.SECONDS)
+    proc.inputStream.bufferedReader().readText()
+} catch (e: IOException) {
+    e.printStackTrace()
+    null
 }
