+++ result/src/commonMain/kotlin/com/github/kittinunf/result/Factory.kt
