+++ result/src/commonMain/kotlin/com/github/kittinunf/result/NoException.kt
