+++ result/src/commonMain/kotlin/com/github/kittinunf/result/Result.kt
