+++ result/src/commonTest/kotlin/com/github/kittinunf/result/FactoryTest.kt
