+++ result/src/commonTest/kotlin/com/github/kittinunf/result/ReadFile.kt
