+++ result/src/commonTest/kotlin/com/github/kittinunf/result/ResultTest.kt
