+++ result/src/commonTest/resources/lorem_long.txt
