+++ result/src/commonTest/resources/lorem_short.txt
