+++ result/src/jvmTest/kotlin/com/github/kittinunf/result/ReadFile.kt
