--- result/src/main/kotlin/com/github/kittinunf/result/NoException.kt
@@ -1,3 +1,0 @@
-package com.github.kittinunf.result
-
-class NoException private constructor() : Exception()
